# backend_srv/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is my service API documentation for scm-simple.",
        "title": "scm-simple API",
        "contact": {
            "name": "API Support",
            "email": "archer.lukman@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "103.235.75.94:8000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/warehouse/create-draft-transfer": {
            "post": {
                "description": "Create New Draft Transaction Warehouse to Warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse_service item_transfer"
                ],
                "summary": "Create New Draft Transaction",
                "parameters": [
                    {
                        "description": "draft transaction transfer payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateDraftWarehouseTransferParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/warehouse/:location_code": {
            "get": {
                "description": "Search location by location code, customer, technician, warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location warehouse customer technician"
                ],
                "summary": "Search location by location code",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "name search by q",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Location code",
                        "name": "location_code",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test-init/{transaction_id}": {
            "get": {
                "description": "Call Test Init after construct code for swagger init",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callany"
                ],
                "summary": "Test init for swagger docs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction Id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CalTestInitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CalTestInitResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "status_approval": {
                            "type": "string"
                        },
                        "transaction_id": {
                            "type": "string"
                        },
                        "warehouse_code": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.CreateDraftWarehouseTransferParams": {
            "type": "object",
            "properties": {
                "location_destination": {
                    "type": "string"
                },
                "location_origin": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}