# backend_srv/internals/api/warehouse.go
package api

import (
	"encoding/json"
	"log"
	"net/http"

	"github.com/julienschmidt/httprouter"
	"scm-simple-luke.com/dir/internals"
	"scm-simple-luke.com/dir/internals/domain"
)

type LocationSearchResponse struct {
	Message string               `json:"message"`
	Data    []domain.LocationRow `json:"data"`
}

// GetLocationByLocationCode godoc
// @Summary      Search location by location code
// @Description  Search location by location code, customer, technician, warehouse
// @Tags         location warehouse customer technician
// @Accept       json
// @Produce      json
// @Param        q              query  string  false  "name search by q"  Format(email)
// @Param        location_code  path  string  false  "Location code"
// @Success      200  {object}  map[string]any
// @Failure      400  {string}  ErrorResponse
// @Router       /api/warehouse/:location_code [get]
func (s *Server) GetLocationByLocationCode(w http.ResponseWriter, req *http.Request, pathParam httprouter.Params) {

	locationCode := pathParam.ByName("location_code")

	whInfo, err := s.service.WarehouseInfo(req.Context(), locationCode)
	if err != nil {
		log.Println(err)

		internals.WriteErrorResponse(w, http.StatusBadRequest, err.Error())
		return
	}

	var resp map[string]any = make(map[string]any)
	resp["Message"] = "lokasi ditemukan"
	resp["Data"] = whInfo

	internals.WriteResponse(w, http.StatusAccepted, resp)
}

type CreateDraftWarehouseTransferParams struct {
	LocationOrigin      string `json:"location_origin"`
	LocationDestination string `json:"location_destination"`
}



// CreateDraftWarehouseTransfer godoc
// @Summary      Create New Draft Transaction
// @Description  Create New Draft Transaction Warehouse to Warehouse
// @Tags         warehouse_service item_transfer
// @Accept       json
// @Produce      json
// @Param		 request        body      CreateDraftWarehouseTransferParams  true  "draft transaction transfer payload"
// @Success      200  {object}  map[string]any
// @Failure      400  {string}  ErrorResponse
// @Router       /api/v1/warehouse/create-draft-transfer [post]
func (s *Server) CreateDraftWarehouseTransfer(w http.ResponseWriter, req *http.Request, pathParam httprouter.Params) {

	ctx := req.Context()

	var ReqBody CreateDraftWarehouseTransferParams
	DataResp := make(map[string]any)

	// read input from Body
	errDecode := json.NewDecoder(req.Body).Decode(&ReqBody)
	if errDecode != nil {
		DataResp["message"] = "error parsing request body"
		internals.WriteResponse(w, http.StatusBadRequest, DataResp)
		return
	}

	NewDraft, err := s.service.WarehouseService.CreateDraftTx(ctx, ReqBody.LocationOrigin, ReqBody.LocationDestination)
	if err != nil {
		DataResp["message"] = err.Error()
		internals.WriteResponse(w, http.StatusConflict, DataResp)
		return
	}
	DataResp["message"] = "create draft berhasil"
	DataResp["data"] = NewDraft
	internals.WriteResponse(w, http.StatusAccepted, DataResp)
	return

}